requirements.txt
tensorflow
opencv-python
numpy
librosa
PySimpleGUI
pygame
Pillow
matplotlib